refactor(deck_stats_panel): split DeckStatsPanel into mixin package
Move widgets/panels/deck_stats_panel.py into a package: frame.py keeps
the class and its inline __init__ UI construction; properties.py holds
the module-level constants, the hypergeometric/colour/curve helpers,
the HTML/CSS/JS builders, and the read-only stat-item getters;
handlers.py owns update_stats, set_card_manager, and clear.
