+++ widgets/panels/deck_stats_panel/__init__.py
@@ -1,0 +1,7 @@
+"""Deck stats panel package."""
+
+from __future__ import annotations
+
+from widgets.panels.deck_stats_panel.frame import DeckStatsPanel
+
+__all__ = ["DeckStatsPanel"]
